+++ internal/cmd/testdata/breaking/one/from/bar/v1/bar.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+
+package bar.v1;
+
+option go_package = "barv1";
+option java_multiple_files = true;
+option java_outer_classname = "BarProto";
+option java_package = "com.bar.v1";
+
+message One {}
