+++ internal/cmd/testdata/breaking/one/from/foo/v1/foo.proto
@@ -7,5 +7,6 @@ option java_multiple_files = true;
 option java_outer_classname = "FooProto";
 option java_package = "com.foo.v1";
 
-// One is a one.
 message One {}
+
+message Two {}
