+++ internal/cmd/testdata/breaking/one/from/prototool.yaml
@@ -3,17 +3,14 @@ lint:
     add:
       - ENUM_FIELD_PREFIXES_EXCEPT_MESSAGE
       - ENUM_ZERO_VALUES_INVALID_EXCEPT_MESSAGE
-      - ENUMS_HAVE_COMMENTS
       - FIELDS_NOT_RESERVED
       - FILE_OPTIONS_EQUAL_GO_PACKAGE_LAST_TWO_SUFFIX
       - GOGO_NOT_IMPORTED
       - IMPORTS_NOT_PUBLIC
       - IMPORTS_NOT_WEAK
       - MESSAGE_FIELDS_NO_JSON_NAME
-      - MESSAGES_HAVE_COMMENTS_EXCEPT_REQUEST_RESPONSE_TYPES
       - PACKAGE_MAJOR_BETA_VERSIONED
       - REQUEST_RESPONSE_NAMES_MATCH_RPC
-      - RPCS_HAVE_COMMENTS
     remove:
       - ENUM_FIELD_PREFIXES
       - ENUM_ZERO_VALUES_INVALID
