+++ internal/cmd/testdata/breaking/one/to/foo/v1/foo.proto
@@ -7,5 +7,4 @@ option java_multiple_files = true;
 option java_outer_classname = "FooProto";
 option java_package = "com.foo.v1";
 
-// One is a one.
 message One {}
